New project with new tests added using MSTest

+++ Clases/IndiceManager.cs
@@ -7,7 +7,7 @@ namespace Clases
 {
     public class IndiceManager
     {
-        public static string TransformarEnLetra(short nota)
+        public static string TransformarEnLetra(int nota)
         {
             string letra;
 
@@ -28,17 +28,17 @@ namespace Clases
 
             return letra;
         }
-        public static decimal TransformarEnValor(short nota)
+        public static float TransformarEnValor(short nota)
         {
-            decimal valor;
+            float valor;
             if (nota >= 90 && nota < 101 )
                 valor = 4;
             else if (nota >= 85 && nota < 90)
-                valor = 3.5M;
+                valor = 3.5f;
             else if (nota >= 80 && nota < 85)
                 valor = 3;
             else if (nota >= 75 && nota < 80)
-                valor = 2.5M;
+                valor = 2.5f;
             else if (nota >= 70 && nota < 75)
                 valor = 2;
             else if (nota >= 60)
@@ -48,33 +48,33 @@ namespace Clases
 
             return valor;
         }
-        public static decimal TransformarEnPuntosHonor(decimal valor, short creditos)
+        public static float TransformarEnPuntosHonor(float valor, short creditos)
         {
             return valor * creditos;
         }
-        public static decimal ObtenerIndice(decimal puntos, short creditos)
+        public static float ObtenerIndice(float puntos, short creditos)
         {
             return puntos / creditos;
         }
-        public static string Honor(decimal indice)
+        public static string Honor(float indice)
         {
             string honor;
 
-            if (indice >= 3.8m)
+            if (indice >= 3.8f)
                 honor = "Summa Cum Laude";
-            else if (indice >= 3.5m)
+            else if (indice >= 3.5f)
                 honor = "Magna Cum Laude";
-            else if (indice >= 3.2m)
+            else if (indice >= 3.2f)
                 honor = "Cum Laude";
             else
                 honor = "Sin honor";
 
             return honor;
         }
-        public static decimal ObtenerIndiceAcumulado(string id)
+        public static float ObtenerIndiceAcumulado(string id)
         {
-            List<decimal> indices = new List<decimal>();
-            decimal valor, totalPuntos = 0;
+            List<float> indices = new List<float>();
+            float valor, totalPuntos = 0;
             short totalCreditos = 0;
 
             List<Trimestre> trimestres = RegistroManager.ObtenerTodosRegistros(id);
@@ -91,9 +91,9 @@ namespace Clases
             }
             return indices.Average(); 
         }
-        public static decimal ObtenerIndiceTrimestre(string id, int trimestre)
+        public static float ObtenerIndiceTrimestre(string id, int trimestre)
         {
-            decimal valor, totalPuntos = 0;
+            float valor, totalPuntos = 0;
             short totalCreditos = 0;
 
             List<Registro> registros = RegistroManager.ObtenerRegistrosTrimestre(id, trimestre);
